requirements.txt
glfw~=2.0.0
numpy~=1.19.5
pyOpenGl
pillow